+++ src/web/AssetManager.php
@@ -13,6 +13,7 @@ use craft\errors\DbConnectException;
 use craft\helpers\App;
 use craft\helpers\Db;
 use craft\helpers\FileHelper;
+use craft\helpers\UrlHelper;
 use yii\caching\TagDependency;
 use yii\db\Exception as DbException;
 
@@ -63,6 +64,8 @@ class AssetManager extends \yii\web\AssetManager
                     $url .= '?v=' . $timestamp;
                 }
             }
+
+            $url = $this->_addBuildIdParam($url);
         }
 
         return $url;
@@ -138,6 +141,24 @@ class AssetManager extends \yii\web\AssetManager
             $url .= '?v=' . $timestamp;
         }
 
-        return [$file, $url];
+        return [$file, $this->_addBuildIdParam($url)];
+    }
+
+    public function getAssetUrl($bundle, $asset, $appendTimestamp = null): string
+    {
+        return $this->_addBuildIdParam(
+            parent::getAssetUrl($bundle, $asset, $appendTimestamp),
+        );
+    }
+
+    private function _addBuildIdParam($url): string
+    {
+        $generalConfig = Craft::$app->getConfig()->getGeneral();
+        if ($generalConfig->buildId) {
+            return UrlHelper::urlWithParams($url, [
+                'buildId' => $generalConfig->buildId,
+            ]);
+        }
+        return $url;
     }
 }